refactor: 사이드바 컴포넌트 이름 변경 및 가맹점 정보 추가
